test: add utils test case
